Move new modules into web_src/js/modules/fomantic/
