+++ web_src/js/features/comp/LabelEdit.ts
@@ -1,5 +1,5 @@
 import {toggleElem} from '../../utils/dom.ts';
-import {showModal} from '../../modules/modal.ts';
+import {showModal} from '../../modules/fomantic/modal.ts';
 import {submitFormFetchAction} from '../common-fetch-action.ts';
 
 function nameHasScope(name: string): boolean {
