+++ web_src/js/features/repo-branch.ts
@@ -1,5 +1,5 @@
 import {toggleElem} from '../utils/dom.ts';
-import {showModal} from '../modules/modal.ts';
+import {showModal} from '../modules/fomantic/modal.ts';
 
 export function initRepoBranchButton() {
   initRepoCreateBranchButton();
