+++ web_src/js/modules/fomantic/modal.ts
@@ -2,6 +2,27 @@ import type {FomanticInitFunction} from '../../types.ts';
 import {queryElems} from '../../utils/dom.ts';
 import {hideToastsFrom} from '../toast.ts';
 
+type ModalOpts = {
+  closable?: boolean;
+  onApprove?: (this: HTMLElement) => boolean | void;
+  onShow?: (this: HTMLElement) => void | Promise<void>;
+  onHide?: (this: HTMLElement) => void;
+  onHidden?: (this: HTMLElement) => void;
+};
+
+// thin wrapper around Fomantic's jQuery modal plugin so callers don't have to touch jQuery or fomanticQuery
+export function showModal(el: Element | null, opts: ModalOpts = {}) {
+  if (!el) return;
+  const $el = $(el);
+  if (Object.keys(opts).length) $el.modal(opts);
+  $el.modal('show');
+}
+
+export function hideModal(el: Element | null) {
+  if (!el) return;
+  $(el).modal('hide');
+}
+
 const fomanticModalFn = $.fn.modal;
 
 // use our own `$.fn.modal` to patch Fomantic's modal module
